Made `add`, `delete`, `edit` and `mark` functions
