+++ src/modules/listUtils.js
@@ -1,4 +1,4 @@
-import { getLocalStorage } from './localStorage.js';
+import { updateLocalStorage, getLocalStorage } from './localStorage.js';
 
 const toDoList = document.querySelector('.todo-list-ul');
 
@@ -39,4 +39,42 @@ export const clearCompleted = (toDoListArray) => {
     task.id = index + 1;
   });
   return toDoListArray;
+};
+
+export const editTask = (e, toDoListArray) => {
+  const clickedTask = e.target.closest('.todo-list-li-text');
+  clickedTask.disabled = false;
+  clickedTask.focus();
+  const taskText = clickedTask.value;
+
+  clickedTask.addEventListener('keypress', (e) => {
+    if (e.key === 'Enter' && clickedTask.value !== '') {
+      const taskIndex = toDoListArray.findIndex((task) => task.task === taskText);
+      toDoListArray[taskIndex].task = clickedTask.value;
+      clickedTask.disabled = true;
+      updateLocalStorage(toDoListArray);
+      renderToDoList(toDoListArray);
+    }
+  });
+};
+
+export const deleteTask = (e, toDoListArray) => {
+  const clickedCross = e.target.closest('.todo-list-li-cross');
+  const clickedTask = clickedCross.previousElementSibling;
+  const taskIndex = toDoListArray.findIndex((task) => task.task === clickedTask.value);
+  toDoListArray.splice(taskIndex, 1);
+  toDoListArray.forEach((task, index) => {
+    task.id = index + 1;
+  });
+  updateLocalStorage(toDoListArray);
+  renderToDoList(toDoListArray);
+};
+
+export const markTask = (e, toDoListArray) => {
+  const clickedCheckbox = e.target.closest('.todo-list-li-checkbox');
+  const clickedTask = clickedCheckbox.nextElementSibling;
+  const taskIndex = toDoListArray.findIndex((task) => task.task === clickedTask.value);
+  toDoListArray[taskIndex].completed = !toDoListArray[taskIndex].completed;
+  updateLocalStorage(toDoListArray);
+  renderToDoList(toDoListArray);
 };
